refactor(refs T30969): move DpResettableInput to components root
