+++ src/components/DpResettableInput/DpResettableInput.stories.mdx
@@ -1,6 +1,6 @@
 # Resettable Input
 
-An input field wich comes with an icon to dismiss the content, if there is any.
+An input field which comes with an icon to dismiss the content, if there is any.
 
 Clicking the dismiss-"x" emits "reset".
